requirements-shared.txt
absl-py==2.1.0
accelerate==0.31.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
blessed==1.20.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
DISTS-pytorch==0.1
einops==0.8.2
exceptiongroup==1.2.0
executing==2.0.1
facexlib==0.3.0
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.4
filterpy==1.4.5
fonttools==4.51.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.3.1
future==1.0.0
gpustat==1.1.1
gradio==4.26.0
gradio_client==0.15.1
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
icecream==2.1.3
idna==3.7
imageio==2.34.1
imgaug==0.4.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
lit==18.1.3
llvmlite==0.42.0
lmdb==1.4.1
lpips==0.1.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.59.1
numpy==2.2.6
nvidia-ml-py==12.535.133
omegaconf==2.3.0
open-clip-torch==3.3.0
openai-clip==1.0.1
opencv-python==4.9.0.80
opencv-python-headless==4.13.0.92
orjson==3.10.0
packaging==24.0
pandas==2.3.3
pillow==10.3.0
platformdirs==4.2.2
protobuf==5.26.1
psutil==5.9.8
pydantic==2.7.0
pydantic_core==2.18.1
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.17.2
pyiqa==0.1.15.post2
pynndescent==0.5.12
pyparsing==3.1.2
pyre-extensions==0.0.23
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-fid==0.3.0
lightning==2.6.1
pytorch-msssim==1.0.0
pytz==2024.1
PyYAML==6.0.3
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.7
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
-e git+https://github.com/simon-schaefer/tar.git@bf28a959fb150ceeadbd9f0bcfc12f3025cf82f4#egg=tar
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2024.5.22
timm==1.0.25
tokenizers==0.22.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch-fidelity==0.4.0
torchmetrics==1.9.0
torchstat==0.0.7
torchsummary==1.5.1
tqdm==4.67.3
transformers==5.3.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
umap-learn==0.5.6
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.2
# Optional on CUDA Linux; MAGC already falls back when xformers is unavailable.
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
